Fix display bug when maps start hidden.
